Convert default LWRP to a custom resource
This could really use some additional refactoring, but testing does show it working
