+++ metadata.rb
@@ -20,4 +20,4 @@ supports 'windows'
 
 source_url 'https://github.com/chef-cookbooks/maven'
 issues_url 'https://github.com/chef-cookbooks/maven/issues'
-chef_version '>= 12.1'
+chef_version '>= 12.5' if respond_to?(:chef_version)
